requirements.txt
rethinkdb
delorean
retrying
praw
bigquery-python